refactoring: naverCrawler.js to crawler.js(because we don't crawl naver)
