+++ src/utils/naverShopApi.js
@@ -3,8 +3,8 @@ const { Card, CardPrice } = require('../models/Card');
 const { Op } = require('sequelize');
 // rarityUtil.js에서 레어도 파싱 함수를 가져옵니다
 const { parseRarity } = require('./rarityUtil');
-// naverCrawler.js에서 나머지 파싱 함수들을 가져옵니다
-const { parseLanguage, parseCondition, extractCardCode, testRarityParsing } = require('./naverCrawler');
+// crawler.js에서 나머지 파싱 함수들을 가져옵니다
+const { parseLanguage, parseCondition, extractCardCode, testRarityParsing } = require('./crawler');
 
 /**
  * 지정된 시간(ms) 동안 실행을 지연시키는 함수
